+++ internal/inventory/types.go
@@ -7,15 +7,15 @@ type ProductID string
 type Product struct {
 	id    ProductID
 	name  string
-	stock int
+	stock uint64
 }
 
 type ReserveItem struct {
 	ProductID ProductID
-	Quantity  int
+	Quantity  uint64
 }
 
-func NewProduct(id ProductID, name string, stock int) *Product {
+func NewProduct(id ProductID, name string, stock uint64) *Product {
 	return &Product{
 		id:    id,
 		name:  name,
@@ -39,11 +39,11 @@ func (p *Product) SetName(name string) {
 	p.name = name
 }
 
-func (p *Product) GetStock() int {
+func (p *Product) GetStock() uint64 {
 	return p.stock
 }
 
-func (p *Product) SetStock(stock int) {
+func (p *Product) SetStock(stock uint64) {
 	p.stock = stock
 }
